# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DipWise API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/portfolios/": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "List Portfolios Route",
        "operationId": "list_portfolios_route_portfolios__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PortfolioResponse"
                  },
                  "type": "array",
                  "title": "Response List Portfolios Route Portfolios  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "portfolios"
        ],
        "summary": "Create Portfolio Route",
        "operationId": "create_portfolio_route_portfolios__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/portfolios/{portfolio_id}/transactions": {
      "post": {
        "tags": [
          "portfolios"
        ],
        "summary": "Add Portfolio Transaction Route",
        "operationId": "add_portfolio_transaction_route_portfolios__portfolio_id__transactions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioTransactionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioTransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "List Portfolio Transactions Route",
        "operationId": "list_portfolio_transactions_route_portfolios__portfolio_id__transactions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PortfolioTransactionResponse"
                  },
                  "title": "Response List Portfolio Transactions Route Portfolios  Portfolio Id  Transactions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/portfolios/{portfolio_id}": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "Get Portfolio Route",
        "operationId": "get_portfolio_route_portfolios__portfolio_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/alerts/events": {
      "get": {
        "tags": [
          "alerts"
        ],
        "summary": "List Alert Events Route",
        "operationId": "list_alert_events_route_alerts_events_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlertEventResponse"
                  },
                  "title": "Response List Alert Events Route Alerts Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/alerts/": {
      "get": {
        "tags": [
          "alerts"
        ],
        "summary": "List Alerts Route",
        "operationId": "list_alerts_route_alerts__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AlertResponse"
                  },
                  "type": "array",
                  "title": "Response List Alerts Route Alerts  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "alerts"
        ],
        "summary": "Create Alert Route",
        "operationId": "create_alert_route_alerts__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlertCreate": {
        "properties": {
          "asset_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Id"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "alert_type": {
            "type": "string",
            "title": "Alert Type",
            "default": "dip_threshold"
          },
          "threshold": {
            "type": "number",
            "title": "Threshold"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "params_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params Json"
          }
        },
        "type": "object",
        "required": [
          "threshold"
        ],
        "title": "AlertCreate"
      },
      "AlertEventResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "alert_id": {
            "type": "integer",
            "title": "Alert Id"
          },
          "triggered_at": {
            "type": "string",
            "format": "date-time",
            "title": "Triggered At"
          },
          "price_at_trigger": {
            "type": "number",
            "title": "Price At Trigger"
          },
          "details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "alert_type": {
            "type": "string",
            "title": "Alert Type"
          },
          "asset_id": {
            "type": "integer",
            "title": "Asset Id"
          },
          "asset_symbol": {
            "type": "string",
            "title": "Asset Symbol"
          }
        },
        "type": "object",
        "required": [
          "id",
          "alert_id",
          "triggered_at",
          "price_at_trigger",
          "details",
          "alert_type",
          "asset_id",
          "asset_symbol"
        ],
        "title": "AlertEventResponse"
      },
      "AlertResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "asset_id": {
            "type": "integer",
            "title": "Asset Id"
          },
          "alert_type": {
            "type": "string",
            "title": "Alert Type"
          },
          "threshold": {
            "type": "number",
            "title": "Threshold"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "params_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params Json"
          }
        },
        "type": "object",
        "required": [
          "id",
          "asset_id",
          "alert_type",
          "threshold",
          "is_active",
          "created_at"
        ],
        "title": "AlertResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PortfolioCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PortfolioCreate"
      },
      "PortfolioDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "positions": {
            "items": {
              "$ref": "#/components/schemas/PositionResponse"
            },
            "type": "array",
            "title": "Positions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "created_at",
          "positions"
        ],
        "title": "PortfolioDetailResponse"
      },
      "PortfolioResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "created_at"
        ],
        "title": "PortfolioResponse"
      },
      "PortfolioTransactionCreate": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "tx_type": {
            "type": "string",
            "enum": [
              "buy",
              "sell"
            ],
            "title": "Tx Type"
          },
          "shares": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Shares"
          },
          "price": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Price"
          },
          "executed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Executed At"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "tx_type",
          "shares",
          "price"
        ],
        "title": "PortfolioTransactionCreate"
      },
      "PortfolioTransactionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "portfolio_id": {
            "type": "integer",
            "title": "Portfolio Id"
          },
          "asset_symbol": {
            "type": "string",
            "title": "Asset Symbol"
          },
          "tx_type": {
            "type": "string",
            "title": "Tx Type"
          },
          "shares": {
            "type": "number",
            "title": "Shares"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "executed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Executed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "portfolio_id",
          "asset_symbol",
          "tx_type",
          "shares",
          "price",
          "executed_at"
        ],
        "title": "PortfolioTransactionResponse"
      },
      "PositionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "asset_symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Symbol"
          },
          "shares": {
            "type": "number",
            "title": "Shares"
          },
          "avg_cost_basis": {
            "type": "number",
            "title": "Avg Cost Basis"
          }
        },
        "type": "object",
        "required": [
          "id",
          "shares",
          "avg_cost_basis"
        ],
        "title": "PositionResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT from POST /auth/login"
      }
    }
  }
}